dailybuild: メモリ不正アクセス予防コード挿入など
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7648 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRScene.cpp
@@ -84,7 +84,7 @@ bool GRScene::AddChildObject(ObjectIf* o){
 			if(ac)
 				sprintf(name, "animation_controller");
 			if(as)
-				sprintf(name, "animation_set%d", animationController->NChildObject()-1);
+				sprintf(name, "animation_set%d", (int)animationController->NChildObject()-1);
 			so->SetName(name);		
 		}
 	}
